% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_words.R
\name{number_to_words}
\alias{number_to_words}
\title{number_to_words}
\usage{
number_to_words(file_path)
}
\arguments{
\item{file_path}{A path to a text file with contents to convert. Required column is the "Test Input" column.}
}
\value{
Prints a tibble with the original text in one column and conversion results in the next column.
}
\description{
This function takes a sentence, checks if there is valid numerical input. If there is, converts the numerical data (up to 11 digits) to words, otherwise returns "number invalid". In the case where the number is valid, but has digits that are greater than 11, it returns "number greater than 11 digits".
}
\examples{
number_to_words("data/test_input_data.txt")
}
